[Fix] common form style 분리
