+++ front/components/account/FindPasswordForm.tsx
@@ -1,7 +1,12 @@
 import { BUTTON_COLOR } from "@utils/constant";
 import { useFormik } from "formik";
 import * as Yup from "yup";
-import { ErrorMessage, FindBtn, Form, Input } from "./FindPasswordForm.style";
+import {
+  InputContainer,
+  Input,
+  SubmitBtn as FindBtn,
+} from "@userContainer/Container.style";
+import { ErrorMessage, Form } from "./FindPasswordForm.style";
 
 interface FindPasswordFormValues {
   email: string;
@@ -26,7 +31,10 @@ function FindPasswordForm() {
         .max(20, "20자 이하로 입력 해 주세요.")
         .required("필수 입력 란입니다."),
       birth: Yup.string()
-        .matches(/^[0-9]{8}$/, "생년월일 8글자를 입력하세요.")
+        .matches(
+          /(19[0-9][0-9]|20\d{2})(0[0-9]|1[0-2])(0[1-9]|[1-2][0-9]|3[0-1])$/,
+          "생년월일을 확인해주세요.",
+        )
         .required("필수 입력 란입니다."),
     }),
     onSubmit: (values) => {
@@ -35,40 +43,46 @@ function FindPasswordForm() {
   });
   return (
     <Form onSubmit={findPasswordFormik.handleSubmit}>
-      <Input
-        id="email"
-        type="email"
-        {...findPasswordFormik.getFieldProps("email")}
-        placeholder="이메일"
-      />
-      {findPasswordFormik.touched.email && findPasswordFormik.errors.email ? (
-        <ErrorMessage>{findPasswordFormik.errors.email}</ErrorMessage>
-      ) : (
-        <ErrorMessage />
-      )}
-      <Input
-        id="email"
-        type="text"
-        {...findPasswordFormik.getFieldProps("name")}
-        placeholder="이름"
-      />
-      {findPasswordFormik.touched.name && findPasswordFormik.errors.name ? (
-        <ErrorMessage>{findPasswordFormik.errors.name}</ErrorMessage>
-      ) : (
-        <ErrorMessage />
-      )}
-      <Input
-        id="birth"
-        type="number"
-        {...findPasswordFormik.getFieldProps("birth")}
-        placeholder="생년월일"
-      />
-      {findPasswordFormik.touched.birth && findPasswordFormik.errors.birth ? (
-        <ErrorMessage>{findPasswordFormik.errors.birth}</ErrorMessage>
-      ) : (
-        <ErrorMessage />
-      )}
-      <FindBtn type="submit" $bgColor={BUTTON_COLOR}>
+      <InputContainer>
+        <Input
+          id="email"
+          type="email"
+          {...findPasswordFormik.getFieldProps("email")}
+          placeholder="이메일"
+        />
+        {findPasswordFormik.touched.email && findPasswordFormik.errors.email ? (
+          <ErrorMessage>{findPasswordFormik.errors.email}</ErrorMessage>
+        ) : (
+          <ErrorMessage />
+        )}
+      </InputContainer>
+      <InputContainer>
+        <Input
+          id="email"
+          type="text"
+          {...findPasswordFormik.getFieldProps("name")}
+          placeholder="이름"
+        />
+        {findPasswordFormik.touched.name && findPasswordFormik.errors.name ? (
+          <ErrorMessage>{findPasswordFormik.errors.name}</ErrorMessage>
+        ) : (
+          <ErrorMessage />
+        )}
+      </InputContainer>
+      <InputContainer>
+        <Input
+          id="birth"
+          type="number"
+          {...findPasswordFormik.getFieldProps("birth")}
+          placeholder="생년월일"
+        />
+        {findPasswordFormik.touched.birth && findPasswordFormik.errors.birth ? (
+          <ErrorMessage>{findPasswordFormik.errors.birth}</ErrorMessage>
+        ) : (
+          <ErrorMessage />
+        )}
+      </InputContainer>
+      <FindBtn type="submit" $btnColor={BUTTON_COLOR}>
         비밀번호 찾기
       </FindBtn>
     </Form>
