+++ front/components/login/LoginForm.style.ts
@@ -17,46 +17,6 @@ export const ContentContainer = styled("div", {
   alignItems: "center",
 });
 
-export const InputContainer = styled("div", {
-  width: "100%",
-  display: "grid",
-  gridTemplateRows: "2.1fr 1fr",
-  justifyItems: "center",
-});
-
-export const Input = styled("input", {
-  width: "65%",
-  height: "3.5rem",
-  border: 0,
-  borderRadius: "25px",
-  fontSize: "1rem",
-  padding: "0 1rem",
-  "::-webkit-input-placeholder": {
-    textAlign: "center",
-    fontFamily: "'Noto Sans KR', sans-serif",
-    color: "#A4A4A4",
-  },
-});
-
-export const ErrorMessage = styled("div", (props: { $txtColor: string }) => ({
-  textAlign: "start",
-  width: "65%",
-  padding: "0.3rem 1rem",
-  color: props.$txtColor,
-  fontSize: "0.9rem",
-}));
-
-export const SubmitBtn = styled("button", (props: { $btnColor: string }) => ({
-  width: "65%",
-  height: "3.5rem",
-  border: 0,
-  borderRadius: "25px",
-  fontSize: "1rem",
-  backgroundColor: props.$btnColor,
-  color: "#fff",
-  fontWeight: "bold",
-}));
-
 export const SubBtnContainer = styled("div", {
   display: "flex",
   justifyContent: "space-evenly",
