+++ front/components/register/UserDataForm.tsx
@@ -48,7 +48,6 @@ function UserDataForm() {
           /(19[0-9][0-9]|20\d{2})(0[0-9]|1[0-2])(0[1-9]|[1-2][0-9]|3[0-1])$/,
           "생년월일을 확인해주세요.",
         )
-        // .matches(/^[0-9]{8}$/, "생년월일 8글자를 입력하세요.")
         .required("필수 입력 란입니다."),
     }),
     onSubmit: (values) => {
